将纯递归无泛型实现移动到独立的文件
